MD RT and Historic

+++ LQPrimaryAPI/RestAPI/Models/Trade.cs
@@ -7,7 +7,7 @@ namespace LatamQuants.PrimaryAPI.Models
     public class Trade
     {
         public double price { get; set; }
-        public int size { get; set; }
+        public double size { get; set; }
         public string datetime { get; set; }
         public string servertime { get; set; }
         public string symbol { get; set; }
